+++ src/callable/primitive/runif.rs
@@ -6,7 +6,7 @@ use crate::ast::*;
 use crate::error::RError;
 use crate::lang::*;
 use crate::callable::core::*;
-use crate::vector::vectors::Vector;
+use crate::vector::Vector;
 
 #[derive(Debug, Clone, PartialEq)]
 #[builtin(sym = "runif")]
