+++ src/vector/coercion/coerce_into.rs
@@ -1,15 +1,13 @@
-use super::{OptionNa, NaAble};
+use crate::vector::types::{OptionNa, NaAble};
 
+/// CoerceInto
+///
+/// A coercion trait often equivalent to [Into]. This allows for slight 
+/// deviations from the way basic rust internals are implemented.
+///
 pub trait CoerceInto<T> {
     fn coerce(self) -> T;
 }
-
-// impl<T> CoerceInto<T> for T
-// {
-//     fn coerce(self) -> T {
-//         self
-//     }
-// }
 
 impl<T, U> CoerceInto<OptionNa<U>> for OptionNa<T> 
 where
@@ -27,8 +25,6 @@ where
 }
 
 // bool
-// impl CoerceInto<OptionNa<bool>> for bool { fn coerce(self) -> OptionNa<bool> { OptionNa(Some(self)) } }
-// impl CoerceInto<OptionNa<i8>> for bool { fn coerce(self) -> OptionNa<i8> { OptionNa(Some(self as i8)) } }
 impl CoerceInto<bool> for bool { fn coerce(self) -> bool { self } }
 impl CoerceInto<i8> for bool { fn coerce(self) -> i8 { self as i8 } }
 impl CoerceInto<i16> for bool { fn coerce(self) -> i16 { self as i16 } }
@@ -38,6 +34,7 @@ impl CoerceInto<i128> for bool { fn coerce(self) -> i128 { self as i128 } }
 impl CoerceInto<f32> for bool { fn coerce(self) -> f32 { self as i32 as f32 } }
 impl CoerceInto<f64> for bool { fn coerce(self) -> f64 { self as i32 as f64 } }
 impl CoerceInto<String> for bool { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for bool 
 where
     Self: CoerceInto<T> 
@@ -48,7 +45,6 @@ where
 }
 
 // i8
-// impl CoerceInto<OptionNa<i8>> for i8 { fn coerce(self) -> OptionNa<i8> { OptionNa(Some(self)) } }
 impl CoerceInto<bool> for i8 { fn coerce(self) -> bool { self != 0 } }
 impl CoerceInto<i8> for i8 { fn coerce(self) -> i8 { self } }
 impl CoerceInto<i16> for i8 { fn coerce(self) -> i16 { self as i16 } }
@@ -58,6 +54,7 @@ impl CoerceInto<i128> for i8 { fn coerce(self) -> i128 { self as i128 } }
 impl CoerceInto<f32> for i8 { fn coerce(self) -> f32 { self as f32 } }
 impl CoerceInto<f64> for i8 { fn coerce(self) -> f64 { self as f64 } }
 impl CoerceInto<String> for i8 { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for i8
 where
     Self: CoerceInto<T> 
@@ -68,7 +65,6 @@ where
 }
 
 // i16
-// impl CoerceInto<OptionNa<i16>> for i16 { fn coerce(self) -> OptionNa<i16> { OptionNa(Some(self)) } }
 impl CoerceInto<bool> for i16 { fn coerce(self) -> bool { self != 0 } }
 impl CoerceInto<i8> for i16 { fn coerce(self) -> i8 { self as i8 } }
 impl CoerceInto<i16> for i16 { fn coerce(self) -> i16 { self } }
@@ -78,6 +74,7 @@ impl CoerceInto<i128> for i16 { fn coerce(self) -> i128 { self as i128 } }
 impl CoerceInto<f32> for i16 { fn coerce(self) -> f32 { self as f32 } }
 impl CoerceInto<f64> for i16 { fn coerce(self) -> f64 { self as f64 } }
 impl CoerceInto<String> for i16 { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for i16
 where
     Self: CoerceInto<T> 
@@ -88,7 +85,6 @@ where
 }
 
 // i32
-// impl CoerceInto<OptionNa<i32>> for i32 { fn coerce(self) -> OptionNa<i32> { OptionNa(Some(self)) } }
 impl CoerceInto<bool> for i32 { fn coerce(self) -> bool { self != 0 } }
 impl CoerceInto<i8> for i32 { fn coerce(self) -> i8 { self as i8 } }
 impl CoerceInto<i16> for i32 { fn coerce(self) -> i16 { self as i16 } }
@@ -98,6 +94,7 @@ impl CoerceInto<i128> for i32 { fn coerce(self) -> i128 { self as i128 } }
 impl CoerceInto<f32> for i32 { fn coerce(self) -> f32 { self as f32 } }
 impl CoerceInto<f64> for i32 { fn coerce(self) -> f64 { self as f64 } }
 impl CoerceInto<String> for i32 { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for i32
 where
     Self: CoerceInto<T> 
@@ -108,7 +105,6 @@ where
 }
 
 // i64
-// impl CoerceInto<OptionNa<i64>> for i64 { fn coerce(self) -> OptionNa<i64> { OptionNa(Some(self)) } }
 impl CoerceInto<bool> for i64 { fn coerce(self) -> bool { self != 0 } }
 impl CoerceInto<i8> for i64 { fn coerce(self) -> i8 { self as i8 } }
 impl CoerceInto<i16> for i64 { fn coerce(self) -> i16 { self as i16 } }
@@ -118,6 +114,7 @@ impl CoerceInto<i128> for i64 { fn coerce(self) -> i128 { self as i128 } }
 impl CoerceInto<f32> for i64 { fn coerce(self) -> f32 { self as f32 } }
 impl CoerceInto<f64> for i64 { fn coerce(self) -> f64 { self as f64 } }
 impl CoerceInto<String> for i64 { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for i64
 where
     Self: CoerceInto<T> 
@@ -128,7 +125,6 @@ where
 }
 
 // i128
-// impl CoerceInto<OptionNa<i128>> for i128 { fn coerce(self) -> OptionNa<i128> { OptionNa(Some(self)) } }
 impl CoerceInto<bool> for i128 { fn coerce(self) -> bool { self != 0 } }
 impl CoerceInto<i8> for i128 { fn coerce(self) -> i8 { self as i8 } }
 impl CoerceInto<i16> for i128 { fn coerce(self) -> i16 { self as i16 } }
@@ -138,6 +134,7 @@ impl CoerceInto<i128> for i128 { fn coerce(self) -> i128 { self } }
 impl CoerceInto<f32> for i128 { fn coerce(self) -> f32 { self as f32 } }
 impl CoerceInto<f64> for i128 { fn coerce(self) -> f64 { self as f64 } }
 impl CoerceInto<String> for i128 { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for i128
 where
     Self: CoerceInto<T> 
@@ -156,6 +153,7 @@ impl CoerceInto<bool> for f32 {
         }
     }
 }
+
 impl CoerceInto<i8> for f32 { fn coerce(self) -> i8 { self as i8 } }
 impl CoerceInto<i16> for f32 { fn coerce(self) -> i16 { self as i16 } }
 impl CoerceInto<i32> for f32 { fn coerce(self) -> i32 { self as i32 } }
@@ -164,6 +162,7 @@ impl CoerceInto<i128> for f32 { fn coerce(self) -> i128 { self as i128 } }
 impl CoerceInto<f32> for f32 { fn coerce(self) -> f32 { self } }
 impl CoerceInto<f64> for f32 { fn coerce(self) -> f64 { self as f64 } }
 impl CoerceInto<String> for f32 { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for f32
 where
     Self: CoerceInto<T>,
@@ -184,7 +183,7 @@ where
         let OptionNa(x) = self;
         match x {
             Some(x) => x.coerce(),
-            None => f32::na(),
+            None => self.na(),
         }
 
     }
@@ -200,6 +199,7 @@ impl CoerceInto<bool> for f64 {
         }
     }
 }
+
 impl CoerceInto<i8> for f64 { fn coerce(self) -> i8 { self as i8 } }
 impl CoerceInto<i16> for f64 { fn coerce(self) -> i16 { self as i16 } }
 impl CoerceInto<i32> for f64 { fn coerce(self) -> i32 { self as i32 } }
@@ -208,6 +208,7 @@ impl CoerceInto<i128> for f64 { fn coerce(self) -> i128 { self as i128 } }
 impl CoerceInto<f32> for f64 { fn coerce(self) -> f32 { self as f32 } }
 impl CoerceInto<f64> for f64 { fn coerce(self) -> f64 { self as f64 } }
 impl CoerceInto<String> for f64 { fn coerce(self) -> String { self.to_string() } }
+
 impl<T> CoerceInto<OptionNa<T>> for f64
 where
     Self: CoerceInto<T>,
@@ -228,15 +229,13 @@ where
         let OptionNa(x) = self;
         match x {
             Some(x) => x.coerce(),
-            None => f64::na(),
+            None => std::f64::NAN,
         }
 
     }
 }
 
-
 // String
-// impl CoerceInto<OptionNa<String>> for String { fn coerce(self) -> OptionNa<String> { OptionNa(Some(self)) } }
 impl CoerceInto<OptionNa<bool>> for String { 
     fn coerce(self) -> OptionNa<bool> { 
         self.to_lowercase()
@@ -244,6 +243,7 @@ impl CoerceInto<OptionNa<bool>> for String {
             .map_or(OptionNa::default(), |i| OptionNa(Some(i)))
     } 
 }
+
 impl CoerceInto<OptionNa<i8>> for String {
     fn coerce(self) -> OptionNa<i8> {
         self.to_lowercase()
@@ -251,6 +251,7 @@ impl CoerceInto<OptionNa<i8>> for String {
             .map_or(OptionNa::default(), |i| OptionNa(Some(i)))
     } 
 }
+
 impl CoerceInto<OptionNa<i16>> for String {
     fn coerce(self) -> OptionNa<i16> {
         self.to_lowercase()
@@ -258,6 +259,7 @@ impl CoerceInto<OptionNa<i16>> for String {
             .map_or(OptionNa::default(), |i| OptionNa(Some(i)))
     } 
 }
+
 impl CoerceInto<OptionNa<i32>> for String {
     fn coerce(self) -> OptionNa<i32> {
         self.to_lowercase()
@@ -265,6 +267,7 @@ impl CoerceInto<OptionNa<i32>> for String {
             .map_or(OptionNa::default(), |i| OptionNa(Some(i)))
     } 
 }
+
 impl CoerceInto<OptionNa<i64>> for String {
     fn coerce(self) -> OptionNa<i64> {
         self.to_lowercase()
@@ -272,6 +275,7 @@ impl CoerceInto<OptionNa<i64>> for String {
             .map_or(OptionNa::default(), |i| OptionNa(Some(i)))
     } 
 }
+
 impl CoerceInto<OptionNa<i128>> for String {
     fn coerce(self) -> OptionNa<i128> {
         self.to_lowercase()
@@ -279,21 +283,25 @@ impl CoerceInto<OptionNa<i128>> for String {
             .map_or(OptionNa::default(), |i| OptionNa(Some(i)))
     } 
 }
+
 impl CoerceInto<f32> for String {
     fn coerce(self) -> f32 {
         self.parse::<f32>().unwrap_or(std::f32::NAN)
     } 
 }
+
 impl CoerceInto<f64> for String {
     fn coerce(self) -> f64 {
         self.parse::<f64>().unwrap_or(std::f64::NAN)
     } 
 }
+
 impl CoerceInto<String> for String {
     fn coerce(self) -> String {
         self
     } 
 }
+
 impl<T> CoerceInto<OptionNa<T>> for String 
 where
     Self: CoerceInto<T>
