+++ src/vector/coercion/common_num.rs
