+++ src/vector/rep/mod.rs
@@ -1,0 +1,6 @@
+//! Vector representations
+mod subsets;
+pub use subsets::*;
+
+mod subset;
+pub use subset::*;
