+++ src/vector/types/na_able.rs
@@ -5,8 +5,7 @@
 /// built-in NaN representation. 
 ///
 pub trait NaAble {
-    type From: Default;
-    fn inner(self) -> Self::From;
+    fn inner(self) -> bool;
     fn is_na(&self) -> bool;
-    fn na() -> Self;
+    fn na(&self) -> bool;
 }
