--- vec/Cargo.toml
@@ -1,4 +1,0 @@
-[package]
-name = "vec"
-version = "0.1.0"
-edition = "2021"
