--- vec/src/coercion/mod.rs
@@ -1,22 +1,0 @@
-mod macros;
-
-mod coerce_into;
-pub use coerce_into::*;
-
-mod option_na;
-pub use option_na::*;
-
-mod na_able;
-pub use na_able::*;
-
-mod into;
-pub use into::*;
-
-mod common_num;
-pub use common_num::*;
-
-mod common_cmp;
-pub use common_cmp::*;
-
-mod map_common;
-pub use map_common::*;
