--- vec/src/types/mod.rs
@@ -1,11 +1,0 @@
-mod numerics;
-pub use numerics::*;
-
-mod integers;
-pub use integers::*;
-
-mod logicals;
-pub use logicals::*;
-
-mod characters;
-pub use characters::*;
